profile 부분 MVC 패턴 적용 연습
